@@ -25,7 +25,7 @@ import java.util.ArrayList;
 
 public class PetInfoActivity extends AppCompatActivity {
 
-    public static final String ARTICLE_KEY = "article";
+    public static final String KEY_ARTICLE = "article";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -33,13 +33,15 @@ public class PetInfoActivity extends AppCompatActivity {
         supportRequestWindowFeature(Window.FEATURE_NO_TITLE);
         setContentView(R.layout.activity_pet_info);
 
-        Pet pet = new Gson().fromJson(getIntent().getStringExtra(PetsActivity.PET_KEY), Pet.class);
+        Pet pet = new Gson().fromJson(getIntent().getStringExtra(PetsActivity.KEY_PET), Pet.class);
 
         ((TextView) findViewById(R.id.pet_name_title_TV)).setText(pet.name);
         ((ImageView) findViewById(R.id.pet_image_IV)).setBackgroundResource(pet.imageResource);
+
         TextView name = findViewById(R.id.pet_name_TV);
         name.setTextColor(Color.rgb(pet.RGB_themeColor[0], pet.RGB_themeColor[1], pet.RGB_themeColor[2]));
         name.setText(pet.name);
+
         ((ListView) findViewById(R.id.pet_articles_LV)).setAdapter(new ArticleAdapter(this, pet.articles));
     }
 
@@ -57,19 +59,22 @@ public class PetInfoActivity extends AppCompatActivity {
             if (convertView == null)
                 convertView = LayoutInflater.from(getContext()).inflate(R.layout.article_item, null);
 
-            GradientDrawable circle = (GradientDrawable) convertView.findViewById(R.id.article_circle_IV).getBackground();
+            // смена цвета кружка
+            ImageView articleCircle = convertView.findViewById(R.id.article_circle_IV);
+            GradientDrawable circle = (GradientDrawable) articleCircle.getBackground();
             circle.setColor(Color.rgb(article.RGB_themeColor[0], article.RGB_themeColor[1], article.RGB_themeColor[2]));
-            ((ImageView) convertView.findViewById(R.id.article_circle_IV)).setBackground(circle);
+            articleCircle.setBackground(circle);
 
             ((TextView) convertView.findViewById(R.id.article_name_TV)).setText(article.name);
 
             final Intent articleActivity = new Intent(PetInfoActivity.this, ArticleActivity.class);
             ((ImageButton) convertView.findViewById(R.id.article_button_IB)).setOnClickListener(new View.OnClickListener() {
+
                 @Override
                 public void onClick(View view) {
                     if (article.imageResource == Pet.DEFAULT_RESOURCE_VALUE) return;
 
-                    articleActivity.putExtra(ARTICLE_KEY, new Gson().toJson(article));
+                    articleActivity.putExtra(KEY_ARTICLE, new Gson().toJson(article));
 
                     startActivity(articleActivity);
                 }
